+++ BrowserJS/Package/Source/DOM/createElement.ts
@@ -1,8 +1,35 @@
+import {
+  Logger,
+  ImproperUsageError,
+  UnexpectedEventError,
+  isUndefined
+} from "@yamato-daiwa/es-extensions";
 import createElements from "./createElements";
-import { Logger, ImproperUsageError } from "@yamato-daiwa/es-extensions";
 
 
-export default function createElement(HTML_Code: string): Element {
+export default function createElement(
+  parametersObject: {
+    HTML_Code: string;
+  }
+): Element;
+
+export default function createElement<SpecificElement extends Element>(
+  parametersObject: {
+    HTML_Code: string;
+    rootElementTypeChecker: (rootElement: Element) => rootElement is SpecificElement;
+  }
+): SpecificElement;
+
+
+export default function createElement<SpecificElement extends Element>(
+  {
+    HTML_Code,
+    rootElementTypeChecker
+  }: {
+    HTML_Code: string;
+    rootElementTypeChecker?: (element: Element) => element is SpecificElement;
+  }
+): Element | SpecificElement {
 
   const elementsCollection: HTMLCollection = createElements(HTML_Code);
 
@@ -17,5 +44,25 @@ export default function createElement(HTML_Code: string): Element {
     });
   }
 
-  return elementsCollection[0];
+
+  if (isUndefined(rootElementTypeChecker)) {
+    return elementsCollection[0];
+  }
+
+
+  const targetElement: Element = elementsCollection[0];
+
+  if (!rootElementTypeChecker(targetElement)) {
+    Logger.throwErrorAndLog({
+      errorInstance: new UnexpectedEventError(
+        "Unable to cast the created element to type specified in 'rootElementTypeChecker' because type check has not " +
+        "passed"
+      ),
+      title: UnexpectedEventError.DEFAULT_TITLE,
+      occurrenceLocation: "createElement(HTML_Code)"
+    });
+  }
+
+
+  return targetElement;
 }