+++ CoreLibrary/Package/Source/TypeGuards/Nullables/isEitherUndefinedOrNull.ts
@@ -1,0 +1,3 @@
+export default function isEitherUndefinedOrNull(rawValue: unknown): rawValue is undefined | null {
+  return typeof rawValue === "undefined" || rawValue === null;
+}