+++ IDEsPlugins/IntelliJ_IDEA/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 # yamato_daiwa-es_extensions-intellij_platform_plugin Changelog
 
+## [0.1.0]
+
+### Added
+
+* [`TimePoint`](https://github.com/TokugawaTakeshi/Yamato-Daiwa-ES-Extensions/blob/master/CoreLibrary/Package/Documentation/DateTime/TimePoint.md)
+  Inserts the example of `TimePoint` class usage
+* [`getArrayElementMatchingWithPredicateIfSuchElementExactlyOne`](https://github.com/TokugawaTakeshi/Yamato-Daiwa-ES-Extensions/blob/master/CoreLibrary/Package/Documentation/Arrays/getArrayElementMatchingWithPredicateIfSuchElementExactlyOne.md)
+  Inserts the example of same-name function usage.
+
+### Changed
+
+* Support for IntelliJ IDEA build **2021.3** 
+
 ## [0.0.2]
 
 ### Added